refactor(settings): extract 4 panes + Field/Segmented atoms (Task #725 Phase A)
Split SettingsDialog.tsx (~771 LoC bag-of-everything) into a thin shell
(~168 LoC: tab nav + Dialog wiring) plus per-pane modules and 2 shared
atoms. Each pane now owns its own state/IPC and is independently testable.

- src/components/settings/AppearancePane.tsx (theme, font size, language, close behaviour)
- src/components/settings/NotificationsPane.tsx (notify mute toggle)
- src/components/settings/UpdatesPane.tsx (version + auto-check + crash reporting)
- src/components/settings/ConnectionPane.tsx (base url, model, discovered models)
- src/components/settings/Field.tsx (shared label/hint atom)
- src/components/settings/Segmented.tsx (shared inline radiogroup atom)

Adds RTL coverage for the two previously untested panes
(NotificationsPane, UpdatesPane) and the two new atoms (Field, Segmented).
Existing connection-pane and appearance tests untouched and still pass.
