+++ src/components/settings/Segmented.tsx
@@ -1,0 +1,46 @@
+import React from 'react';
+import { cn } from '../../lib/cn';
+
+// Inline radiogroup used across Settings panes for small enum choices
+// (theme, language, close behaviour, etc.). Generic over the value type so
+// each callsite keeps its narrowed string-literal union.
+export function Segmented<T extends string>({
+  value,
+  onChange,
+  options,
+}: {
+  value: T;
+  onChange: (v: T) => void;
+  options: { value: T; label: string }[];
+}) {
+  return (
+    <div
+      className={cn(
+        'inline-flex h-7 items-center rounded-sm border border-border-default',
+        'bg-bg-elevated p-0.5 gap-0.5'
+      )}
+      role="radiogroup"
+    >
+      {options.map((o) => {
+        const active = o.value === value;
+        return (
+          <button
+            key={o.value}
+            role="radio"
+            aria-checked={active}
+            onClick={() => onChange(o.value)}
+            className={cn(
+              'h-6 px-2.5 text-meta rounded-[3px] transition-[background-color,color,box-shadow] duration-150 ease-out',
+              'outline-none focus-ring',
+              active
+                ? 'bg-bg-app text-fg-primary font-medium shadow-[inset_0_0_0_1px_var(--color-border-default)]'
+                : 'text-fg-secondary hover:text-fg-primary hover:bg-bg-hover'
+            )}
+          >
+            {o.label}
+          </button>
+        );
+      })}
+    </div>
+  );
+}
